# docs/contracts/portal-api-v1.4.1/docs/web-portal-api/openapi.yaml
...
openapi: "3.1.0"
info:
  title: SSPOS web portal API
  version: "1.4.0"
  summary: "The contract between SSPOS 3 tills and the web portal: sync, licensing, devices, onboarding, migration."
  description: "Contract major version 1 (X-SSPOS-Contract: 1), document revision 1.4 (package v1.4, 2026-09-29). The prose contract is docs/web-portal-api.md; where this file and a sample disagree, the sample wins. Schemas are JSON Schema 2020-12 referenced by relative path (schemas/ is generated from the till's C# types; licensing/schemas/ is the section 17 contract). Every reply may gain new fields; clients ignore fields they do not know. The till calls nine operations: sync hello / push / pull, licence activate / validate / redeem, devices/deactivate, cloud/migrate and its complete; devices/activate is kept for reference only (deprecated - not called). Every POST except a normal (delta) push carries an Idempotency-Key, one per logical request: a retry repeats the key and the first attempt's body byte for byte (section 17.11 rule 5)."
  license:
    name: Proprietary - SSPOS
    identifier: LicenseRef-SSPOS-Proprietary
servers:
  - url: "https://{portalHost}"
    description: The portal. HTTPS only (TLS 1.2+).
    variables:
      portalHost:
        default: portal.example.co.uk
...
security:
  - branchKey: []
paths:
...
  /api/v1/sync/push:
    post:
      tags:
        - Sync
      operationId: syncPush
      summary: "Till to portal: a batch of changed rows (sections 5 and 7)"
      description: "Body is gzip (Content-Encoding: gzip); accept an uncompressed body too (test tools). Delta mode keys rows by (branchId, seq). Initial mode (X-SSPOS-Sync-Mode: initial + X-SSPOS-Upload-Id, section 17.8) keys them by (uploadId, seq). Idempotent by (entity, entityId, version): storing a triple twice changes nothing and gives the same reply. Every time in the body is UTC with Z (section 6.1). The reply carries receivedAt (revision 1.3): when the portal durably stored the batch, UTC with Z, by the portal clock; store it on every row as received_at and answer a retried batch with the receivedAt its rows were first stored with. Revision 1.4: each entry is the row's latest state - version = the row's rowVersion when the batch was built (payload.rowVersion); several entries of one row in a batch carry the same triple, payload and op (store once, count all as accepted), and a resend may carry a newer state under a newer version (section 7); Setting and RolePermission rows arrive as keyed rows (entityId derived from their key, version = seq; key them by the payload - section 10.3); Customer.balance and Customer.points are the till's cache - sum the CustomerTransaction ledger instead (section 10.1); User rows never carry remoteApprovalSecret (section 10.7); a shop's head-office orders come back with the shop's own status and receipts (section 10.6); a SyncConflict row (a clash on the till) carries hubChange, your pulled envelope as JSON text (section 8). A normal push carries no Idempotency-Key; each history-upload batch carries one (the same for the same bytes until answered)."
      parameters:
        - $ref: "#/components/parameters/Contract"
        - $ref: "#/components/parameters/AppVersion"
        - $ref: "#/components/parameters/StoreProtocol"
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/BranchId"
        - $ref: "#/components/parameters/RegisterId"
        - $ref: "#/components/parameters/ContentEncoding"
        - $ref: "#/components/parameters/SyncMode"
        - $ref: "#/components/parameters/UploadId"
        - $ref: "#/components/parameters/IdempotencyKeyOptional"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "./schemas/push-request.schema.json"
            examples:
              leeds:
                summary: Leeds till 1 (delta)
                externalValue: "./samples/push-request.json"
              leedsTill2:
                summary: A Leeds till 2 sale, sent by till 1
                externalValue: "./samples/push-request.second-till.json"
              bradford:
                summary: Bradford
                externalValue: "./samples/push-request.second-branch.json"
              settings:
                summary: "Leeds: a shared setting changed, a permission granted and one taken away (section 10.3)"
                externalValue: "./samples/push-request.settings.json"
              initial:
                summary: First rows of an initial (migration) upload
                value:
                  - seq: 1
                    entity: VatRate
                    entityId: "01K5T0Q8C4000000000000V001"
                    op: I
                    version: 1
                    companyId: "01K5M1GR8T000000000000C009"
                    branchId: ""
                    registerId: ""
                    at: "2026-09-01T08:00:00Z"
                    payload:
                      name: Standard
                      code: S
                      percentage: 20.0
                      treatment: apply
                      effectiveFrom: "2011-01-04"
                      effectiveTo: null
                      isDefault: true
                      isZeroValued: false
                      id: "01K5T0Q8C4000000000000V001"
                      companyId: "01K5M1GR8T000000000000C009"
                      createdAt: "2026-09-01T08:00:00Z"
                      updatedAt: "2026-09-01T08:00:00Z"
                      rowVersion: 1
                      deletedAt: null
                      isDeleted: false
                      domainEvents: []
                    key: "VatRate:01K5T0Q8C4000000000000V001:1"
                  - seq: 2
                    entity: VatRate
                    entityId: "01K5T0Q8C4000000000000V002"
                    op: I
                    version: 1
                    companyId: "01K5M1GR8T000000000000C009"
                    branchId: ""
                    registerId: ""
                    at: "2026-09-01T08:00:00Z"
                    payload:
                      name: Zero
                      code: Z
                      percentage: 0.0
                      treatment: apply
                      effectiveFrom: "1973-04-01"
                      effectiveTo: null
                      isDefault: false
                      isZeroValued: true
                      id: "01K5T0Q8C4000000000000V002"
                      companyId: "01K5M1GR8T000000000000C009"
                      createdAt: "2026-09-01T08:00:00Z"
                      updatedAt: "2026-09-01T08:00:00Z"
                      rowVersion: 1
                      deletedAt: null
                      isDeleted: false
                      domainEvents: []
                    key: "VatRate:01K5T0Q8C4000000000000V002:1"
      responses:
        "200":
          description: OK
          headers:
            X-SSPOS-Contract:
              $ref: "#/components/headers/Contract"
          content:
            "application/json":
              schema:
                $ref: "./schemas/push-reply.schema.json"
              examples:
                ack:
                  summary: Acknowledged
                  value:
                    acknowledgedSeq: 18239
                    accepted: 9
                    receivedAt: "2026-09-23T09:42:13.207Z"
        "400":
          $ref: "#/components/responses/E400"
        "401":
          $ref: "#/components/responses/E401"
        "403":
          $ref: "#/components/responses/E403"
        "404":
          $ref: "#/components/responses/E404"
        "409":
          $ref: "#/components/responses/E409"
        "413":
          $ref: "#/components/responses/E413"
        "422":
          $ref: "#/components/responses/E422"
        "426":
          $ref: "#/components/responses/E426"
        "429":
          $ref: "#/components/responses/E429"
        "500":
          $ref: "#/components/responses/E500"
        "503":
          $ref: "#/components/responses/E503"
...
  /api/v1/cloud/migrate:
    post:
      tags:
        - Migration
      operationId: migrateToCloud
      summary: "Link a branch that never synced and open its history upload - Connect with a sync key (section 17.8)"
      description: "Sent by the branch's main till when the owner presses Connect with a sync key and the branch has never sent anything (route A of SIMPLE-SETUP.md) - on any licence (a local key, the built-in trial or a portal licence: for a branch you already know answer 200 adopted with the same ids) - and when a local-licence shop enters a portal licence key. No Authorization header: activationCode (the sync key, which must also work as the branch's sync bearer, or the portal licence key) is the credential. Idempotent per (activationCode, installId): a retry returns the same reply and uploadId; the till repeats the Idempotency-Key and the first body (tillClockUtc, registers[].lastSeenUtc included) until you answer. After 200 the till says hello with apiKey, uploads its history with POST /api/v1/sync/push in initial mode, then calls /api/v1/cloud/migrate/complete. A wrong, expired or used sync key may be named sync_key.not_found / sync_key.expired / sync_key.used or activation.code_* (section 17.12)."
      security: []
      parameters:
        - $ref: "#/components/parameters/Contract"
        - $ref: "#/components/parameters/AppVersion"
        - $ref: "#/components/parameters/StoreProtocol"
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/BranchId"
        - $ref: "#/components/parameters/RegisterId"
        - $ref: "#/components/parameters/InstallId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "./licensing/schemas/migrate-request.schema.json"
            examples:
              hydePark:
                summary: Hyde Park, local licence
                value:
                  activationCode: SSP-2WQA-HN4D-R7KM
                  installId: "01K5M1GR8T000000000000H009"
                  installCode: "0M01-TFG6"
                  deviceName: HYDEPARK-TILL
                  appVersion: "3.0.412"
                  os:
                    name: Windows 11 Pro
                    version: "10.0.26200"
                    architecture: x64
                  tillClockUtc: "2026-09-26T11:29:58Z"
                  approverKids:
                    - k76b44696
                  trustedKids:
                    - k13799fa4
                    - k76b44696
                  localLicenceToken: SSPOS1.eyJ2IjoxLCJraWQiOiJrNzZiNDQ2OTYiLCJsaWNlbmNlSWQiOiIwMUs1TTFHUjhUMDAwMDAwMDAwMDAwWTAwMiIsImtpbmQiOiJmdWxsIiwic291cmNlIjoibG9jYWwiLCJpc3N1ZXIiOiJTU1BPUyBMaWNlbmNlIEdlbmVyYXRvciAtIE93bmVyIFBDIiwiY29tcGFueUlkIjoiIiwiYnJhbmNoSWQiOiIiLCJidXNpbmVzc05hbWUiOiJIeWRlIFBhcmsgTmV3cyBMdGQiLCJicmFuY2hOYW1lIjoiSHlkZSBQYXJrIiwiaW5zdGFsbENvZGUiOiIwTTAxLVRGRzYiLCJtYXhSZWdpc3RlcnMiOjIsImZlYXR1cmVzIjpbIndoYXRzYXBwIl0sImlzc3VlZEF0IjoiMjAyNi0wMy0zMVQxMDowMDowMFoiLCJ2YWxpZEZyb20iOiIyMDI2LTAzLTMxVDAwOjAwOjAwWiIsImV4cGlyZXNBdCI6IjIwMjctMDMtMzFUMjM6NTk6NTlaIiwibm90ZXMiOiJQYWlkIGJ5IGJhbmsgdHJhbnNmZXIsIGludm9pY2UgMTA0MiJ9.c2IlqHOT1WguKa1H6aTEvKWKBiIoejoTubfayzeZhqx-g6apnMwPfM-R1vIbE_DAspKZrY289Bqm7fPWCtOtCg
                  localTrialEndsAt: null
                  company:
                    id: "01K5M1GR8T000000000000C009"
                    name: Hyde Park News Ltd
                  branch:
                    id: "01K5M1GR8T000000000000B009"
                    name: Hyde Park
                    code: HPK
                  registers:
                    - registerId: "01K5M1GR8T000000000000R009"
                      code: "01"
                      name: Till 1
                      isMain: true
                      appVersion: "3.0.412"
                      installId: "01K5M1GR8T000000000000H009"
                      deviceName: HYDEPARK-TILL
                      lastSeenUtc: "2026-09-26T11:29:58Z"
                      isActive: true
                  data:
                    totalRows: 159306
                    rowCounts:
                      VatRate: 4
                      Department: 12
                      Category: 58
                      Product: 3120
                      ProductBarcode: 3388
                      Customer: 41
                      User: 5
                      Register: 1
                      BranchProduct: 3120
                      Shift: 176
                      Sale: 15840
                      SaleLine: 41210
                      SalePayment: 16102
                      SaleVat: 22017
                      StockMovement: 44916
                      ZReport: 176
                      AuditLog: 9120
                    firstSaleAtUtc: "2026-04-01T07:02:11Z"
                    lastSaleAtUtc: "2026-09-26T11:20:40Z"
                    snapshotChangeLogSeq: 212806
      responses:
        "200":
          description: OK
          headers:
            X-SSPOS-Contract:
              $ref: "#/components/headers/Contract"
          content:
            "application/json":
              schema:
                $ref: "./licensing/schemas/migrate-reply.schema.json"
              examples:
                adopted:
                  summary: "New account: portal adopts the till's ids"
                  value:
                    companyId: "01K5M1GR8T000000000000C009"
                    branchId: "01K5M1GR8T000000000000B009"
                    registers:
                      - registerId: "01K5M1GR8T000000000000R009"
                        seat: allowed
                    hubUrl: "https://portal.example.co.uk"
                    apiKey: sspos_br_SAMPLE_c4d2a9e0f7b3418e95aa
                    licenceToken: SSPOS1.eyJ2IjoxLCJraWQiOiJrMTM3OTlmYTQiLCJsaWNlbmNlSWQiOiIwMUs1VDBROEM0MDAwMDAwMDAwMDAwWTAwMyIsImtpbmQiOiJmdWxsIiwic291cmNlIjoicG9ydGFsIiwiaXNzdWVyIjoiU1NQT1MgUG9ydGFsIiwiY29tcGFueUlkIjoiMDFLNU0xR1I4VDAwMDAwMDAwMDAwMEMwMDkiLCJicmFuY2hJZCI6IjAxSzVNMUdSOFQwMDAwMDAwMDAwMDBCMDA5IiwiYnVzaW5lc3NOYW1lIjoiSHlkZSBQYXJrIE5ld3MgTHRkIiwiYnJhbmNoTmFtZSI6Ikh5ZGUgUGFyayIsIm1heFJlZ2lzdGVycyI6MiwiZmVhdHVyZXMiOlsid2hhdHNhcHAiXSwiaXNzdWVkQXQiOiIyMDI2LTA5LTI2VDExOjMwOjAwWiIsInZhbGlkRnJvbSI6IjIwMjYtMDktMjZUMTE6MzA6MDBaIiwiZXhwaXJlc0F0IjoiMjAyNy0wMy0zMVQyMzo1OTo1OVoiLCJvbmxpbmVDaGVjayI6eyJyZXF1aXJlZCI6dHJ1ZSwiaW50ZXJ2YWxIb3VycyI6MjQsImdyYWNlRGF5cyI6MTR9LCJub3RlcyI6Ik1vdmVkIGZyb20gbG9jYWwgbGljZW5jZSAwMUs1TTFHUjhUMDAwMDAwMDAwMDAwWTAwMiAoMTg2IGRheXMgY2FycmllZCBvdmVyKSJ9.VIVGeb_kRwrle0a1DPgXGplq4Snr61SOB4qewKN_lJ9Z3Bs88nTudB1K1CHQCkOybamSeW0l1oBjshiLNNq3DQ
                    licence:
                      licenceId: "01K5T0Q8C4000000000000Y003"
                      kind: full
                      source: portal
                      expiresAt: "2027-03-31T23:59:59Z"
                      maxRegisters: 2
                      features:
                        - whatsapp
                      limits: null
                    carriedOverDays: 186
                    upload:
                      uploadId: "01K5TA00000000000000000G01"
                      mode: initial
                      maxRowsPerBatch: 5000
                      resumeFromSeq: 0
                    settingsBootstrap: null
                    portalTimeUtc: "2026-09-26T11:30:00Z"
                    nextCheckAfterSeconds: 86400
                    messages:
                      - id: "01K5T0Q8C4000000000000M007"
                        level: info
                        title: Uploading your history
                        text: Your sales history is being copied to the portal in the background. Keep trading as normal.
                        showFromUtc: null
                        showUntilUtc: null
                        dismissible: true
                        link: null
                    idMapping:
                      company:
                        localId: "01K5M1GR8T000000000000C009"
                        portalId: "01K5M1GR8T000000000000C009"
                        action: adopted
                      branch:
                        localId: "01K5M1GR8T000000000000B009"
                        portalId: "01K5M1GR8T000000000000B009"
                        action: adopted
                aliased:
                  summary: "Existing account: portal keeps its company id"
                  value:
                    companyId: "01K5M1GR8T000000000000C009"
                    branchId: "01K5M1GR8T000000000000B009"
                    registers:
                      - registerId: "01K5M1GR8T000000000000R009"
                        seat: allowed
                    hubUrl: "https://portal.example.co.uk"
                    apiKey: sspos_br_SAMPLE_c4d2a9e0f7b3418e95aa
                    licenceToken: SSPOS1.eyJ2IjoxLCJraWQiOiJrMTM3OTlmYTQiLCJsaWNlbmNlSWQiOiIwMUs1VDBROEM0MDAwMDAwMDAwMDAwWTAwMyIsImtpbmQiOiJmdWxsIiwic291cmNlIjoicG9ydGFsIiwiaXNzdWVyIjoiU1NQT1MgUG9ydGFsIiwiY29tcGFueUlkIjoiMDFLNU0xR1I4VDAwMDAwMDAwMDAwMEMwMDkiLCJicmFuY2hJZCI6IjAxSzVNMUdSOFQwMDAwMDAwMDAwMDBCMDA5IiwiYnVzaW5lc3NOYW1lIjoiSHlkZSBQYXJrIE5ld3MgTHRkIiwiYnJhbmNoTmFtZSI6Ikh5ZGUgUGFyayIsIm1heFJlZ2lzdGVycyI6MiwiZmVhdHVyZXMiOlsid2hhdHNhcHAiXSwiaXNzdWVkQXQiOiIyMDI2LTA5LTI2VDExOjMwOjAwWiIsInZhbGlkRnJvbSI6IjIwMjYtMDktMjZUMTE6MzA6MDBaIiwiZXhwaXJlc0F0IjoiMjAyNy0wMy0zMVQyMzo1OTo1OVoiLCJvbmxpbmVDaGVjayI6eyJyZXF1aXJlZCI6dHJ1ZSwiaW50ZXJ2YWxIb3VycyI6MjQsImdyYWNlRGF5cyI6MTR9LCJub3RlcyI6Ik1vdmVkIGZyb20gbG9jYWwgbGljZW5jZSAwMUs1TTFHUjhUMDAwMDAwMDAwMDAwWTAwMiAoMTg2IGRheXMgY2FycmllZCBvdmVyKSJ9.VIVGeb_kRwrle0a1DPgXGplq4Snr61SOB4qewKN_lJ9Z3Bs88nTudB1K1CHQCkOybamSeW0l1oBjshiLNNq3DQ
                    licence:
                      licenceId: "01K5T0Q8C4000000000000Y003"
                      kind: full
                      source: portal
                      expiresAt: "2027-03-31T23:59:59Z"
                      maxRegisters: 2
                      features:
                        - whatsapp
                      limits: null
                    carriedOverDays: 186
                    upload:
                      uploadId: "01K5TA00000000000000000G01"
                      mode: initial
                      maxRowsPerBatch: 5000
                      resumeFromSeq: 0
                    settingsBootstrap: null
                    portalTimeUtc: "2026-09-26T11:30:00Z"
                    nextCheckAfterSeconds: 86400
                    messages:
                      - id: "01K5T0Q8C4000000000000M007"
                        level: info
                        title: Uploading your history
                        text: Your sales history is being copied to the portal in the background. Keep trading as normal.
                        showFromUtc: null
                        showUntilUtc: null
                        dismissible: true
                        link: null
                    idMapping:
                      company:
                        localId: "01K5M1GR8T000000000000C009"
                        portalId: "01K5T0Q8C4000000000000C001"
                        action: aliased
                      branch:
                        localId: "01K5M1GR8T000000000000B009"
                        portalId: "01K5M1GR8T000000000000B009"
                        action: adopted
        "400":
          $ref: "#/components/responses/E400"
        "403":
          $ref: "#/components/responses/E403"
        "404":
          $ref: "#/components/responses/E404"
        "409":
          $ref: "#/components/responses/E409"
        "410":
          $ref: "#/components/responses/E410"
        "422":
          $ref: "#/components/responses/E422"
        "426":
          $ref: "#/components/responses/E426"
        "429":
          $ref: "#/components/responses/E429"
        "500":
          $ref: "#/components/responses/E500"
        "503":
          $ref: "#/components/responses/E503"
  /api/v1/cloud/migrate/complete:
    post:
      tags:
        - Migration
      operationId: completeMigration
      summary: Close the initial upload (section 17.8)
      description: "Sent by the branch's main till after the last upload batch was acknowledged, with Authorization: Bearer <apiKey from the migrate reply>. Idempotent per uploadId and by Idempotency-Key (kept until you answer). complete: the till resumes normal push after pushResumesAfterSeq and pulls from since = 0. incomplete: it resends from resumeFromSeq, then calls complete again with a new key."
      parameters:
        - $ref: "#/components/parameters/Contract"
        - $ref: "#/components/parameters/AppVersion"
        - $ref: "#/components/parameters/StoreProtocol"
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/BranchId"
        - $ref: "#/components/parameters/RegisterId"
        - $ref: "#/components/parameters/InstallId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "./licensing/schemas/migrate-complete-request.schema.json"
            examples:
              done:
                summary: All rows sent
                value:
                  uploadId: "01K5TA00000000000000000G01"
                  totalRows: 159306
                  highestSeq: 159306
                  rowCounts:
                    VatRate: 4
                    Department: 12
                    Category: 58
                    Product: 3120
                    ProductBarcode: 3388
                    Customer: 41
                    User: 5
                    Register: 1
                    BranchProduct: 3120
                    Shift: 176
                    Sale: 15840
                    SaleLine: 41210
                    SalePayment: 16102
                    SaleVat: 22017
                    StockMovement: 44916
                    ZReport: 176
                    AuditLog: 9120
                  snapshotChangeLogSeq: 212806
      responses:
        "200":
          description: OK
          headers:
            X-SSPOS-Contract:
              $ref: "#/components/headers/Contract"
          content:
            "application/json":
              schema:
                $ref: "./licensing/schemas/migrate-complete-reply.schema.json"
              examples:
                complete:
                  summary: Counts match
                  value:
                    uploadId: "01K5TA00000000000000000G01"
                    status: complete
                    received: 159306
                    resumeFromSeq: 159306
                    missing: []
                    pushResumesAfterSeq: 212806
                    portalTimeUtc: "2026-09-26T14:02:10Z"
                    messages: []
                incomplete:
                  summary: "Some rows missing: resend"
                  value:
                    uploadId: "01K5TA00000000000000000G01"
                    status: incomplete
                    received: 154306
                    resumeFromSeq: 140000
                    missing:
                      - entity: StockMovement
                        expected: 44916
                        received: 39916
                    pushResumesAfterSeq: 212806
                    portalTimeUtc: "2026-09-26T14:02:10Z"
                    messages: []
        "400":
          $ref: "#/components/responses/E400"
        "401":
          $ref: "#/components/responses/E401"
        "403":
          $ref: "#/components/responses/E403"
        "404":
          $ref: "#/components/responses/E404"
        "409":
          $ref: "#/components/responses/E409"
        "422":
          $ref: "#/components/responses/E422"
        "426":
          $ref: "#/components/responses/E426"
        "429":
          $ref: "#/components/responses/E429"
        "500":
          $ref: "#/components/responses/E500"
        "503":
          $ref: "#/components/responses/E503"
components:
  securitySchemes:
    branchKey:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: "The branch's sync key (sync.hub_api_key): typed by the owner at Connect, or returned by the portal as apiKey in licence/activate, licence/validate or cloud/migrate. One key per branch. Required on sync/* and cloud/migrate/complete; sent on licence/* and devices/deactivate only when the till holds one."
  parameters:
    Contract:
      name: X-SSPOS-Contract
      in: header
      required: true
      description: Contract major version the till speaks. 1 today. See section 17.11.
      schema:
        type: string
        pattern: "^[0-9]+$"
      example: "1"
    AppVersion:
      name: X-SSPOS-App-Version
      in: header
      required: true
      description: Till version.
      schema:
        $ref: "./licensing/schemas/common.schema.json#/$defs/appVersion"
      example: "3.0.412"
    StoreProtocol:
      name: X-SSPOS-Store-Protocol
      in: header
      required: true
      description: "The till's LAN store protocol version (src/SSPOS.Server/Contracts/StoreProtocol.cs), sent on the sync and migration calls (not on licence or device calls). 3 since 2026-09-28: 2 when a second till's work began to be written on the main till as that till, 3 when the main till's staff list (and, as optional calls within 3, its shop-wide settings and shop prices) began to be copied to every till (docs/web-portal-api.md section 2 point 2); 1 before. Informational: accept any number."
      schema:
        type: string
        pattern: "^[0-9]+$"
      example: "3"
    CompanyId:
      name: X-SSPOS-Company-Id
      in: header
      required: true
      description: Company of the calling till.
      schema:
        $ref: "./licensing/schemas/common.schema.json#/$defs/ulid"
      example: "01K5T0Q8C4000000000000C001"
    BranchId:
      name: X-SSPOS-Branch-Id
      in: header
      required: true
      description: "Branch of the calling till; must be the API key's branch."
      schema:
        $ref: "./licensing/schemas/common.schema.json#/$defs/ulid"
      example: "01K5T0Q8C4000000000000B001"
    RegisterId:
      name: X-SSPOS-Register-Id
      in: header
      required: true
      description: "The calling till: on sync and migration calls always the branch's main till (the sender); on licence and device calls the till licensing or releasing itself."
      schema:
        $ref: "./licensing/schemas/common.schema.json#/$defs/ulid"
      example: "01K5T0Q8C4000000000000R001"
...
    IdempotencyKeyOptional:
      name: Idempotency-Key
      in: header
      required: false
      description: "Absent on a normal (delta) push - its rows are the idempotency, by (entity, entityId, version) and (branchId, seq). Present on every history-upload batch (X-SSPOS-Sync-Mode: initial): one key per batch, repeated with the same bytes until the portal answers (section 17.11 rule 5)."
      schema:
        type: string
        minLength: 8
        maxLength: 64
    ContentEncoding:
      name: Content-Encoding
      in: header
      required: false
      description: The till always sends gzip.
      schema:
        type: string
        enum:
          - gzip
      example: gzip
    SyncMode:
      name: X-SSPOS-Sync-Mode
      in: header
      required: false
      description: "delta (default) or initial (a migration upload). Unknown = 400."
      schema:
        type: string
        enum:
          - delta
          - initial
        default: delta
      example: delta
    UploadId:
      name: X-SSPOS-Upload-Id
      in: header
      required: false
      description: "Required when X-SSPOS-Sync-Mode is initial: upload.uploadId from the migrate reply."
      schema:
        $ref: "./licensing/schemas/common.schema.json#/$defs/ulid"
...
  headers:
...
    Contract:
      description: Contract major version of this reply.
      schema:
        type: string
        pattern: "^[0-9]+$"
      example: "1"
...
  responses:
    E400:
      description: "Bad request. Codes: `request.invalid` (*). Full list: licensing/samples/error-codes.json."
      headers:
        X-SSPOS-Contract:
          $ref: "#/components/headers/Contract"
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/ErrorReply"
          examples:
            request-invalid:
              summary: request.invalid
              value:
                code: request.invalid
                message: Body or query does not match the schema (missing field, bad ULID, bad date).
                traceId: req-8a21f0
                retryAfterSeconds: null
                rejectedKey: null
                details:
                  field: name of the first bad field
    E401:
      description: "Unauthorised. Codes: `auth.invalid_key` (sync/*, licence/validate, licence/redeem, devices/deactivate, cloud/migrate/complete); `auth.key_revoked` (sync/*, licence/validate, licence/redeem, devices/deactivate, cloud/migrate/complete). Full list: licensing/samples/error-codes.json."
      headers:
        X-SSPOS-Contract:
          $ref: "#/components/headers/Contract"
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/ErrorReply"
          examples:
            sync-401:
              summary: samples/error-reply.401.json
              value:
                code: auth.invalid_key
                message: "The portal API key is not recognised. Check Settings > Hub sync > Portal API key."
                traceId: req-7c1f9a
                retryAfterSeconds: null
                rejectedKey: null
    E403:
      description: "Forbidden. Codes: `auth.wrong_branch` (sync/*, licence/validate, licence/redeem, devices/deactivate); `device.not_main_till` (licence/validate, licence/redeem, devices/deactivate, cloud/migrate/complete); `licence.seat_limit` (devices/activate, cloud/migrate, licence/activate); `licence.not_active` (devices/activate, cloud/migrate, licence/activate); `key.not_for_this_branch` (licence/redeem); `key.not_allowed` (licence/redeem). Full list: licensing/samples/error-codes.json."
      headers:
        X-SSPOS-Contract:
          $ref: "#/components/headers/Contract"
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/ErrorReply"
          examples:
            seat-limit:
              summary: error.seat-limit.403
              value:
                code: licence.seat_limit
                message: Your plan allows 2 tills and 2 are in use. Add a till to your plan on the portal.
                traceId: req-8a21f0
                retryAfterSeconds: null
                rejectedKey: null
                details:
                  maxRegisters: 2
                  registersInUse: 2
    E404:
      description: "Not found. Codes: `activation.code_not_found` (devices/activate, cloud/migrate); `device.not_found` (devices/deactivate); `key.not_found` (licence/redeem, licence/activate, licence/validate); `migrate.upload_not_found` (sync/push (initial), cloud/migrate/complete); `sync_key.not_found` (cloud/migrate). Full list: licensing/samples/error-codes.json."
      headers:
        X-SSPOS-Contract:
          $ref: "#/components/headers/Contract"
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/ErrorReply"
          examples:
            activation-code-not-found:
              summary: error.activation-code-not-found.404
              value:
                code: activation.code_not_found
                message: That activation code is not recognised. Check the code in your welcome email.
                traceId: req-8a21f0
                retryAfterSeconds: null
                rejectedKey: null
                details: null
...
    E413:
      description: "Payload too large. Codes: `batch.too_large` (sync/push). Full list: licensing/samples/error-codes.json."
      headers:
        X-SSPOS-Contract:
          $ref: "#/components/headers/Contract"
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/ErrorReply"
          examples:
            batch-too_large:
              summary: batch.too_large
              value:
                code: batch.too_large
                message: More rows or bytes than accepted.
                traceId: req-8a21f0
                retryAfterSeconds: null
                rejectedKey: null
                details:
                  maxRows: 5000
...
    E426:
      description: "Upgrade required. Codes: `app.update_required` (*). Full list: licensing/samples/error-codes.json."
      headers:
        X-SSPOS-Contract:
          $ref: "#/components/headers/Contract"
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/ErrorReply"
          examples:
            update-required:
              summary: error.update-required.426
              value:
                code: app.update_required
                message: "This till needs updating before it can link. Settings > Updates > Check now."
                traceId: req-8a21f0
                retryAfterSeconds: null
                rejectedKey: null
                details:
                  minimumAppVersion: "3.0.400"
    E429:
      description: "Too many requests. Codes: `rate.limited` (*); `activation.too_many_attempts` (devices/activate, cloud/migrate, licence/activate). Full list: licensing/samples/error-codes.json."
      headers:
        X-SSPOS-Contract:
          $ref: "#/components/headers/Contract"
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/ErrorReply"
          examples:
            activation-too-many-attempts:
              summary: error.activation-too-many-attempts.429
              value:
                code: activation.too_many_attempts
                message: Too many wrong codes. Try again in 15 minutes.
                traceId: req-8a21f0
                retryAfterSeconds: 900
                rejectedKey: null
                details: null
...